Move TailwindCSS and main.js files to 'docs' directory
The commit moves the TailwindCSS and main.js files from the 'src' directory to the 'docs' directory. Consequently, the script and link urls in package.json and index.html have been updated to reflect the new file locations. This rearrangement aims to provide better organization and clarity of project structure.