Update styles, scripts, and cleanup project files
This commit includes the addition of new CSS classes in output.css for additional styling and layout control, and optimizes animations in the main.js file. Updates to the index.html improve the structure and mobile view of the content. Config options were also added to tailwind.config.js and redundant entries were cleaned up in .idea/.gitignore and .idea/jsLibraryMappings.xml. This elevates the quality of the codebase and enhances the user experience.